pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "This is the test task I received during interview process"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.0.2"
pydantic-settings = "^2.2.1"
psycopg2-binary = "^2.9.9"
gunicorn = "^21.2.0"
django-cte = "^1.3.2"
django-import-export = "^3.3.7"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
